refactor(repo): remove redundant webhookerror for generic UrlboxError
